PythonProject18/Crypto News/Azure_function_updater/requirements.txt
azure-functions
motor
feedparser
langchain-openai
python-dotenv
